MAINT: Refactored "task status" to "task state"
