+++ topobank/frontend/analysis/TaskStateRow.vue
